+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## HEAD (unreleased)
 
+- Feat (`ngx-select`): `withMargin` input when set to false will display ngx-select marginless.
+- Feat (`ngx-select`, `ngx-input`): `withHint` input when set to false will not display a block element for the hint.
+
 ## 51.0.0-alpha.1 (2025-10-21)
 
 - Fix (`ngx-checkbox`): Removed duplicate output emission